Create Container for BookList component
